STEP-11: close metadata section numbers to 1 through 11
Parsed and native metadata sections carry a closed enum. Wire admission rejects other numbers and section-size matching is exhaustive.

Committed with --no-verify: the scoped gate runs in the lane's verification step and after merge.

+++ crates/cadmpeg-codec-inventor/src/validate.rs
@@ -1616,7 +1616,7 @@ fn validate_segments(data: &NativeData, findings: &mut Vec<Finding>) {
             sections
                 .entry(record.token.as_str())
                 .or_default()
-                .insert(record.number);
+                .insert(u8::from(record.number));
             sections
         },
     );
